Changes to Training & Qualifications summary page for a staff member.
- Removed 'Add a record' button
- Created 'Add a training record' button
- Created 'Add a qualification record' button

When adding a training record, if no training courses
have been previously added, clicking the button will
navigate to the 'select category' page.

If courses have already been added, it will
navigate to a new page to be developed in a
separate ticket. Logic has been put in
place ready for this but it is not
functional yet as there is no means yet to
add training courses from the frontend.
For now, in these cases, the button will navigate
to the 'select category' page.

The 'Add a qualification record' button
navigates to the existing 'add a qualification
record' page.

e2e tests have been updated.